move to scripts
