+++ scripts/train_tokenizer.py
@@ -1,4 +1,5 @@
 import datasets
+from loguru import logger
 import numpy as np
 from transformers import AutoTokenizer
